+++ src/app/(auth)/(with-navbar)/register/page.tsx
@@ -18,6 +18,8 @@ const RegisterPage: NextPage = () => {
         phoneNumber: '',
         password: '',
         confirmPassword: '',
+        identification: '',
+        identificationType: 'CC',
         checkTerms: false,
     });